backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
motor>=3.3.2

# Environment
python-dotenv>=1.0.0

# Groq API
groq>=0.4.2

# Document Processing
pypdf>=4.0.1

# Vector Search & Embeddings
sentence-transformers>=2.3.1
faiss-cpu>=1.9.0
numpy>=1.26.0

# CORS
starlette>=0.35.1

# Validation
pydantic>=2.5.3
